chore: semantic structure
