+++ packages/upgrade/src/lib/ast/modifiers.ts
@@ -1,27 +1,3 @@
-import type {
-  ASTPath,
-  Identifier,
-  JSCodeshift,
-  MemberExpression,
-} from 'jscodeshift';
-
-/**
- * Walks the path upwards to look for the closest parent
- * of the mentioned type
- */
-export function findParentOfType<TPath>(
-  path: ASTPath<unknown>,
-  type: JSCodeshift['AnyTypeAnnotation'],
-): ASTPath<unknown> | false {
-  if (!type.check(path.node)) {
-    return findParentOfType(path.parentPath, type);
-  }
-  if (!path.parent) {
-    return false;
-  }
-  return path as ASTPath<TPath>;
-}
-
 /**
  * Replaces the identifier for the root path key
  * of a member expression
